full dir data

+++ include/packr/entry.h
@@ -1,6 +1,7 @@
 #pragma once
 #define NAME_LEN_MAX 4096
 #include <packr/types.h>
+#include <packr/ops.h>
 
 struct pack_header {
     char dirname[NAME_LEN_MAX];
@@ -15,10 +16,9 @@ struct pack_header {
     u64 mod_time; // last modification time
     u64 sc_time;  // last status change time   u64 total_file_count;
     u16 dirname_length;
-    u8 mode; // permissions
+    u16 mode; // permissions
 };
 pack_header* pack_header_init();
-
 
 struct file_entry {
     char filename[NAME_LEN_MAX];
@@ -27,7 +27,7 @@ struct file_entry {
     u64 mod_time; // last modification time
     u64 sc_time;  // last status change time
     u16 filename_length;
-    u8 mode; // file size
+    u16 mode; // file size
 };
 
 struct dir_entry {
@@ -40,5 +40,5 @@ struct dir_entry {
     u64 mod_time; // last modification time
     u64 sc_time;  // last status change time
     u16 dirname_length;
-    u8 mode; // permissions
+    u16 mode; // permissions
 };
